# emmaus_ui/src/app/services/models/page-response-impact-story-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ImpactStoryResponse } from '../models/impact-story-response';
export interface PageResponseImpactStoryResponse {
  content?: Array<ImpactStoryResponse>;
  empty?: boolean;
  first?: boolean;
  last?: boolean;
  number?: number;
  numberOfElements?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}