# emmaus_ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/programs/update/{public-id}": {
            "put": {
                "tags": [
                    "programs"
                ],
                "operationId": "updateProgram",
                "parameters": [
                    {
                        "name": "public-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProgramRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/impact-stories/update/{public-id}": {
            "put": {
                "tags": [
                    "impact-stories"
                ],
                "operationId": "updateImpactStory",
                "parameters": [
                    {
                        "name": "public-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateImpactStoryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/programs/create": {
            "post": {
                "tags": [
                    "programs"
                ],
                "operationId": "createProgram",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProgramRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/impact-stories/create": {
            "post": {
                "tags": [
                    "impact-stories"
                ],
                "operationId": "createImpactStory",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateImpactStoryRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/programs/pages": {
            "get": {
                "tags": [
                    "programs"
                ],
                "operationId": "getPagesOfPrograms",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "public-id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseProgramResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/impact-stories/statistics": {
            "get": {
                "tags": [
                    "impact-stories"
                ],
                "operationId": "getImpactStatistics",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "last-days",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImpactStatisticsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/impact-stories/pages": {
            "get": {
                "tags": [
                    "impact-stories"
                ],
                "operationId": "getPagesOfImpactStories",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "public-id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseImpactStoryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/programs/{public-id}": {
            "delete": {
                "tags": [
                    "programs"
                ],
                "operationId": "deleteProgram",
                "parameters": [
                    {
                        "name": "public-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/impact-stories/{public-id}": {
            "delete": {
                "tags": [
                    "impact-stories"
                ],
                "operationId": "deleteImpactStory",
                "parameters": [
                    {
                        "name": "public-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateProgramRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "metricNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "metricLabel": {
                        "type": "string"
                    },
                    "logoUrl": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "CreateImpactStoryRequest": {
                "type": "object",
                "properties": {
                    "programId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "beneficiaryName": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "gender": {
                        "type": "string",
                        "enum": [
                            "FEMALE",
                            "MALE"
                        ]
                    },
                    "location": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "videoUrl": {
                        "type": "string"
                    },
                    "shortQuote": {
                        "type": "string"
                    },
                    "fullStory": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "PageResponseProgramResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProgramResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
            "ProgramResponse": {
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "metricNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "metricLabel": {
                        "type": "string"
                    },
                    "logoUrl": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ImpactStatisticsResponse": {
                "type": "object",
                "properties": {
                    "totalImpacts": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "childImpacts": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "womenImpacts": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "elderlyImpacts": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "specialNeedsImpacts": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ImpactStoryResponse": {
                "type": "object",
                "properties": {
                    "publicId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "programId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "program": {
                        "type": "string"
                    },
                    "beneficiaryName": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "videoUrl": {
                        "type": "string"
                    },
                    "shortQuote": {
                        "type": "string"
                    },
                    "fullStory": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    }
                }
            },
            "PageResponseImpactStoryResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ImpactStoryResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "last": {
                        "type": "boolean"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "empty": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}